refactor: Убрал нормализацию на странице со списком фоток
Раньше в ядре был стейт, куда планировалось загружать только данные фоток, причём для всех страниц, а не только для списка фоток. Теперь стейт перенесён в модуль со списком фоток и в нём хранятся только те фотки, которые отображаются на странице. Благодаря этому код стал понятнее, но стейт будет занимать больше места.
